Move cached_downloader to unix specific file
